renamed and moved skywire ssh
