+++ ci_scripts/run-internal-tests.sh
@@ -19,16 +19,16 @@ go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m 
 
 go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/internal/therealproxy -run TestProxy >> ./logs/internal/TestProxy.log
 
-go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/internal/therealssh  -run TestListAuthoriser >> ./logs/internal/TestListAuthoriser.log
-go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/internal/therealssh  -run TestFileAuthoriser >> ./logs/internal/TestFileAuthoriser.log
-go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/internal/therealssh  -run TestChannelServe >> ./logs/internal/TestChannelServe.log
-go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/internal/therealssh  -run TestChannelSendWrite >> ./logs/internal/TestChannelSendWrite.log
-go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/internal/therealssh  -run TestChannelRead >> ./logs/internal/TestChannelRead.log
-go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/internal/therealssh  -run TestChannelRequest >> ./logs/internal/TestChannelRequest.log
-go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/internal/therealssh  -run TestChannelServeSocket >> ./logs/internal/TestChannelServeSocket.log
-go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/internal/therealssh  -run TestClientOpenChannel >> ./logs/internal/TestClientOpenChannel.log
-go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/internal/therealssh  -run TestClientHandleResponse >> ./logs/internal/TestClientHandleResponse.log
-go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/internal/therealssh  -run TestClientHandleData >> ./logs/internal/TestClientHandleData.log
-go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/internal/therealssh  -run TestServerOpenChannel >> ./logs/internal/TestServerOpenChannel.log
-go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/internal/therealssh  -run TestServerHandleRequest >> ./logs/internal/TestServerHandleRequest.log
-go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/internal/therealssh  -run TestServerHandleData >> ./logs/internal/TestServerHandleData.log
+go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/skywire-messenger-ssh/internal/therealssh  -run TestListAuthoriser >> ./logs/internal/TestListAuthoriser.log
+go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/skywire-messenger-ssh/internal/therealssh  -run TestFileAuthoriser >> ./logs/internal/TestFileAuthoriser.log
+go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/skywire-messenger-ssh/internal/therealssh  -run TestChannelServe >> ./logs/internal/TestChannelServe.log
+go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/skywire-messenger-ssh/internal/therealssh  -run TestChannelSendWrite >> ./logs/internal/TestChannelSendWrite.log
+go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/skywire-messenger-ssh/internal/therealssh  -run TestChannelRead >> ./logs/internal/TestChannelRead.log
+go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/skywire-messenger-ssh/internal/therealssh  -run TestChannelRequest >> ./logs/internal/TestChannelRequest.log
+go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/skywire-messenger-ssh/internal/therealssh  -run TestChannelServeSocket >> ./logs/internal/TestChannelServeSocket.log
+go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/skywire-messenger-ssh/internal/therealssh  -run TestClientOpenChannel >> ./logs/internal/TestClientOpenChannel.log
+go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/skywire-messenger-ssh/internal/therealssh  -run TestClientHandleResponse >> ./logs/internal/TestClientHandleResponse.log
+go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/skywire-messenger-ssh/internal/therealssh  -run TestClientHandleData >> ./logs/internal/TestClientHandleData.log
+go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/skywire-messenger-ssh/internal/therealssh  -run TestServerOpenChannel >> ./logs/internal/TestServerOpenChannel.log
+go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/skywire-messenger-ssh/internal/therealssh  -run TestServerHandleRequest >> ./logs/internal/TestServerHandleRequest.log
+go clean -testcache &>/dev/null || go test -race -tags no_ci -cover -timeout=5m github.com/skycoin/skywire/skywire-messenger-ssh/internal/therealssh  -run TestServerHandleData >> ./logs/internal/TestServerHandleData.log
