--- cmd/therealssh-cli/therealssh-cli.go
@@ -1,12 +1,0 @@
-/*
-CLI for SSH app
-*/
-package main
-
-import (
-	"github.com/skycoin/skywire/cmd/therealssh-cli/commands"
-)
-
-func main() {
-	commands.Execute()
-}
