+++ skywire-messenger-ssh/cmd/skywire-messenger-ssh-cli/README.md
@@ -1,6 +1,6 @@
 # CLI for SSH app
 
-`therealssh-cli` implements PTY related operations for SSH app.
+`skywire-messenger-ssh-cli` implements PTY related operations for SSH app.
 
 It connects to SSH client's RPC server, configures terminal for raw pty
 data and handles pty data exchange.
