+++ skywire-messenger-ssh/cmd/skywire-messenger-ssh-cli/skywire-messenger-ssh-cli.go
@@ -1,0 +1,12 @@
+/*
+CLI for SSH app
+*/
+package main
+
+import (
+	"github.com/skycoin/skywire/skywire-messenger-ssh/cmd/skywire-messenger-ssh-cli/commands"
+)
+
+func main() {
+	commands.Execute()
+}
