+++ skywire-messenger-ssh/cmd/skywire-messenger-ssh-client/README.md
@@ -1,0 +1,8 @@
+# Skywire SSH client app
+
+`skywire-messenger-ssh-client` app implements client for the SSH app. 
+
+It starts RCP interface for `skywire-messenger-ssh-cli` and handles incoming requests to
+the remote node via `skywire` connection.
+
+Please check docs for `skywire-messenger-ssh-server` app for further instructions.
