+++ skywire-messenger-ssh/internal/therealssh/shell.go
