+++ skywire-messenger-ssh/internal/therealssh/shell_darwin.go
